backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
torch==2.8.0
torchvision==0.23.0
Pillow
numpy==1.26.4

